add custom node compilation flags

+++ src/custom_nodes/east_ocr/east_ocr.cpp
@@ -139,7 +139,7 @@ int execute(const struct CustomNodeTensor* inputs, int inputsCount, struct Custo
     NODE_ASSERT(confidenceThreshold >= 0 && confidenceThreshold <= 1.0, "confidence threshold must be in 0-1 range");
     float overlapThreshold = get_float_parameter("overlap_threshold", params, paramsCount, 0.3);
     NODE_ASSERT(overlapThreshold >= 0 && overlapThreshold <= 1.0, "non max suppression filtering overlap threshold must be in 0-1 range");
-    int maxOutputBatch = get_int_parameter("max_output_batch", params, paramsCount, 100);
+    uint64_t maxOutputBatch = get_int_parameter("max_output_batch", params, paramsCount, 100);
     NODE_ASSERT(maxOutputBatch > 0, "max output batch must be larger than 0");
     bool debugMode = get_string_parameter("debug", params, paramsCount) == "true";
     float xCornerAdjustment = get_float_parameter("x_corner_adjustment", params, paramsCount, 0.12);
